+++ service/comprehend/errors.go
@@ -111,11 +111,9 @@ const (
 	// ErrCodeUnsupportedLanguageException for service response error code
 	// "UnsupportedLanguageException".
 	//
-	// Amazon Comprehend can't process the language of the input text. For all custom
-	// entity recognition APIs (such as CreateEntityRecognizer), only English, Spanish,
-	// French, Italian, German, or Portuguese are accepted. For most other APIs,
-	// such as those for Custom Classification, Amazon Comprehend accepts text in
-	// all supported languages. For a list of supported languages, see supported-languages.
+	// Amazon Comprehend can't process the language of the input text. For custom
+	// entity recognition APIs, only English, Spanish, French, Italian, German,
+	// or Portuguese are accepted. For a list of supported languages, see supported-languages.
 	ErrCodeUnsupportedLanguageException = "UnsupportedLanguageException"
 )